Module merino::docs::overview

High level overview of Merino

This project is structured as a Cargo Workspace that contains one crate for each broad area of behavior for Merino. This structure is advantageous because the crates can be handled either individually or as a group. When compiling, each crate can be compiled in parallel, where dependencies allow, and when running tests, each test suite can be run separately or together. This also provides an advantage if we choose to re-use any of these crates in other projects, or if we publish the crates to Crates.io.

This is a brief overview of the crates found in the repository. For more details, see the specific crate docs.

merino

This is the main Merino application, and one of the binary crates in the repository. It brings together and configures the other crates to create a production-like environment for Firefox suggest.

merino-settings

This defines and documents the settings of the application. These settings should be initialized by one of the binary crates, and passed into the other crates to configure them.

merino-web

This crate provides and HTTP API to access Merino, including providing observability into the running of the application via that API.

merino-suggest

This is a domain crate that defines the data model and traits needed to provide suggestions to Firefox.

merino-adm

This crate provides integration with the AdMarketplace APIs, and implements the traits from merino-suggest.

merino-integration-tests

This crate is a separate test system. It works much like merino, in that it brings together the other crates to produce a complete Merino environment. However, this binary crate produces an application that exercise the service as a whole, instead of providing a server to manual test against.